+++ web3lib/src/commonMain/kotlin/com/sonsofcrypto/web3lib/provider/ProviderExtensions.kt
@@ -1,11 +1,11 @@
 package com.sonsofcrypto.web3lib.provider
 
-import com.sonsofcrypto.web3lib.provider.model.DataHexString
+import com.sonsofcrypto.web3lib.provider.model.DataHexStr
 import com.sonsofcrypto.web3lib.provider.model.TransactionRequest
 import com.sonsofcrypto.web3lib.types.Address
 import com.sonsofcrypto.web3lib.types.AddressHexString
 
 @Throws(Throwable::class)
-suspend fun Provider.call(to: AddressHexString, data: DataHexString): DataHexString {
+suspend fun Provider.call(to: AddressHexString, data: DataHexStr): DataHexStr {
     return call(TransactionRequest(to = Address.HexString(to), data = data))
 }