+++ web3lib/src/commonMain/kotlin/com/sonsofcrypto/web3lib/provider/model/QuantityDataHexString.kt
@@ -6,26 +6,26 @@ import com.sonsofcrypto.web3lib.utils.extensions.toHexString
 import kotlinx.serialization.json.JsonPrimitive
 
 /** SEE: https://ethereum.org/en/developers/docs/apis/json-rpc/#quantities-encoding */
-typealias QuantityHexString = String
+typealias QntHexStr = String
 
-fun QuantityHexString.toIntQnt(): Int = this.stripPrefixPad().toInt(16)
-fun QuantityHexString.toUIntQnt(): UInt = this.stripPrefixPad().toUInt(16)
-fun QuantityHexString.toLongQnt(): Long = this.stripPrefixPad().toLong(16)
-fun QuantityHexString.toULongQnt(): ULong = this.stripPrefixPad().toULong(16)
-fun QuantityHexString.toBigIntQnt(): BigInt = BigInt.from(this.stripPrefixPad(), 16)
-fun QuantityHexString.toByteArrayQnt(): ByteArray = this.stripPrefixPad().hexStringToByteArray()
+fun QntHexStr.toIntQnt(): Int = this.stripPrefixPad().toInt(16)
+fun QntHexStr.toUIntQnt(): UInt = this.stripPrefixPad().toUInt(16)
+fun QntHexStr.toLongQnt(): Long = this.stripPrefixPad().toLong(16)
+fun QntHexStr.toULongQnt(): ULong = this.stripPrefixPad().toULong(16)
+fun QntHexStr.toBigIntQnt(): BigInt = BigInt.from(this.stripPrefixPad(), 16)
+fun QntHexStr.toByteArrayQnt(): ByteArray = this.stripPrefixPad().hexStringToByteArray()
 
-fun QuantityHexString(int: Int): QuantityHexString = int.toString(16).addPrefixPad()
-fun QuantityHexString(uint: UInt): QuantityHexString = uint.toString(16).addPrefixPad()
-fun QuantityHexString(long: Long): QuantityHexString = long.toString(16).addPrefixPad()
-fun QuantityHexString(ulong: ULong): QuantityHexString = ulong.toString(16).addPrefixPad()
-fun QuantityHexString(bigInt: BigInt): QuantityHexString = bigInt.toHexString().addPrefixPad()
+fun QntHexStr(int: Int): QntHexStr = int.toString(16).addPrefixPad()
+fun QntHexStr(uint: UInt): QntHexStr = uint.toString(16).addPrefixPad()
+fun QntHexStr(long: Long): QntHexStr = long.toString(16).addPrefixPad()
+fun QntHexStr(ulong: ULong): QntHexStr = ulong.toString(16).addPrefixPad()
+fun QntHexStr(bigInt: BigInt): QntHexStr = bigInt.toHexString().addPrefixPad()
 
 /**
  * QuantityHexString (part of JsonRpc) should have no leading zeros
  * SEE: https://ethereum.org/en/developers/docs/apis/json-rpc/#quantities-encoding
  */
-fun QuantityHexString.stripPrefixPad(): String {
+fun QntHexStr.stripPrefixPad(): String {
     return (if (this.length % 2 != 0) "0"  else "") + this.substring(2)
 }
 
@@ -33,24 +33,24 @@ fun QuantityHexString.stripPrefixPad(): String {
  * QuantityHexString (part of JsonRpc) should have no leading zeros
  * SEE: https://ethereum.org/en/developers/docs/apis/json-rpc/#quantities-encoding
  */
-fun QuantityHexString.addPrefixPad(): String {
+fun QntHexStr.addPrefixPad(): String {
     return (
         if (this.first().toString() == "0" && this.length > 1) this.substring(1)
         else this
     ).addPrefix()
 }
 
-fun QuantityHexString.addPrefix(): String {
+fun QntHexStr.addPrefix(): String {
     return if (length == 1 || (length > 1 && substring(0, 2) != "0x")) "0x" + this else this
 }
 
-fun QuantityHexString.jsonPrimitive(): JsonPrimitive = JsonPrimitive(this)
+fun QntHexStr.jsonPrimitive(): JsonPrimitive = JsonPrimitive(this)
 
 /** SEE: https://ethereum.org/en/developers/docs/apis/json-rpc/#unformatted-data-encoding */
-typealias DataHexString = String
+typealias DataHexStr = String
 
 @OptIn(kotlin.ExperimentalUnsignedTypes::class)
-fun DataHexString(byteArray: ByteArray): DataHexString = byteArray.toHexString(true)
+fun DataHexStr(byteArray: ByteArray): DataHexStr = byteArray.toHexString(true)
 
-fun DataHexString.toByteArrayData(): ByteArray = hexStringToByteArray()
-fun DataHexString.toBigIntData(): BigInt = BigInt.from(this, 16)
+fun DataHexStr.toByteArrayData(): ByteArray = hexStringToByteArray()
+fun DataHexStr.toBigIntData(): BigInt = BigInt.from(this, 16)