+++ CHANGELOG.md
@@ -14,6 +14,16 @@ head over to [sonsofcrypto.com](https://sonsofcrypto.com/).
 
 # Release Notes 
 
+### v1.2.0 #49
+- Signer refactor
+- VoidSigner
+- KeySigner
+- ExtKeySigner
+- MnemonicSigner
+- Legacy transaction support
+- EIP-2718 support
+- EIP-2930 support
+
 ### v1.1.4 #48
 - Wallet type indicator icon
 - Swipe up / down to show wallets search bar